# Conflicts:
#   BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/service/BoardServiceImpl.java

+++ BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/dto/board/BoardLikeDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class BoardLikeDto {
     Boolean isLike;
-    Integer like;
+    Long like;
 }
